@@ -18,7 +18,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=3.0.2 <4.0.0"
 
 dependencies:
   flutter:
@@ -26,8 +26,8 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  flutter_svg: ^1.1.0 # it help us to use SVG in our app
+  cupertino_icons: ^1.0.5
+  flutter_svg: ^2.0.7 # it help us to use SVG in our app
 
 dev_dependencies:
   flutter_test:
