@@ -4,15 +4,15 @@ class Product {
   final String image, title, description;
   final int price, size, id;
   final Color color;
-  Product({
-    this.id,
-    this.image,
-    this.title,
-    this.price,
-    this.description,
-    this.size,
-    this.color,
-  });
+
+  Product(
+      {required this.image,
+      required this.title,
+      required this.description,
+      required this.price,
+      required this.size,
+      required this.id,
+      required this.color});
 }
 
 List<Product> products = [
